requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.1
Flask-RESTful==0.3.2
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
aniso8601==0.92
argparse==1.2.1
itsdangerous==0.24
pymongo==2.8
pytz==2015.2
six==1.9.0
wsgiref==0.1.2
PyYAML==3.11
simplejson==3.8.2
googlefinance==0.7
